refactor: dto에서 유효성 검사 및 형변환 로직 분리
