+++ src/main/java/lotto/dto/request/PaymentRequest.java
@@ -1,11 +1,7 @@
 package lotto.dto.request;
 
-import lotto.dto.ConvertUtil;
 
 public record PaymentRequest(
     int payment
 ) {
-    public static PaymentRequest from(String input) {
-        return new PaymentRequest(ConvertUtil.convertToNumber(input));
-    }
 }
